+++ CHANGELOG.md
@@ -6,5 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+- Support multiple root nodes, an empty root will be synthesized for them.
+
 ## [0.1.0] - 2025-04-04
 Initial release